backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
python-dotenv
pillow
opencv-python-headless
numpy